+++ lib/controllers/session_controller.dart
@@ -1,4 +1,5 @@
 import 'dart:async';
+import 'dart:convert';
 import 'dart:developer';
 
 import 'package:agora_rtc_engine/agora_rtc_engine.dart';
@@ -17,6 +18,7 @@ import 'package:agora_uikit/models/agora_settings.dart';
 import 'package:agora_uikit/models/agora_user.dart';
 import 'package:agora_uikit/src/enums.dart';
 import 'package:flutter/material.dart';
+import 'package:http/http.dart' as http;
 import 'package:permission_handler/permission_handler.dart';
 
 class SessionController extends ValueNotifier<AgoraSettings> {
@@ -73,7 +75,6 @@ class SessionController extends ValueNotifier<AgoraSettings> {
     await value.engine!
         .initialize(RtcEngineContext(appId: value.connectionData!.appId));
     log("SDK initialized: ${value.engine}", level: Level.error.value);
-
     // Getting SDK versions and assigning them
     SDKBuildInfo? rtcVersion = await value.engine?.getVersion();
     AgoraVersions.staticRTM = await AgoraRtmClient.getSdkVersion();
@@ -164,6 +165,8 @@ class SessionController extends ValueNotifier<AgoraSettings> {
 
   /// Function to join the video call.
   Future<void> joinVideoChannel() async {
+    if (value.layoutType == Layout.oneToOne && value.users.length == 1) return;
+
     // [generatedRtmId] is the unique ID for a user generated using the timestamp in milliseconds.
     value = value.copyWith(
       generatedRtmId: value.connectionData!.rtmUid ??
@@ -280,5 +283,39 @@ class SessionController extends ValueNotifier<AgoraSettings> {
 
   void updateLayoutType({required Layout updatedLayout}) {
     value = value.copyWith(layoutType: updatedLayout);
+  }
+
+  Future<void> startCloudRecording(
+      {required AgoraConnectionData connectionData}) async {
+    final response = await http.post(
+      Uri.parse('${connectionData.cloudRecordingUrl}/api/start/call'),
+      body: {"channel": connectionData.channelName},
+    );
+
+    if (response.statusCode == 200) {
+      log('Recording Started', level: Level.warning.value);
+      value = value.copyWith(sid: jsonDecode(response.body)['data']['sid']);
+    } else {
+      log('Couldn\'t start the recording : ${response.statusCode}',
+          level: Level.error.value);
+    }
+  }
+
+  Future<void> stopCloudRecording(
+      {required AgoraConnectionData connectionData}) async {
+    final response = await http.post(
+      Uri.parse('${connectionData.cloudRecordingUrl}/api/stop/call'),
+      body: {
+        "channel": connectionData.channelName,
+        "sid": value.sid,
+      },
+    );
+
+    if (response.statusCode == 200) {
+      log('Recording Ended', level: Level.warning.value);
+    } else {
+      log('Couldn\'t end the recording : ${response.statusCode}',
+          level: Level.error.value);
+    }
   }
 }