+++ lib/models/agora_connection_data.dart
@@ -21,6 +21,9 @@ class AgoraConnectionData {
   /// (Optional) Link to the deployed token server. The UIKit automatically generates the token after a fixed interval. Have a look at this guide to learn how to set up your [token server](https://github.com/AgoraIO-Community/agora-token-service)
   final String? tokenUrl;
 
+  /// (Optional) Link to the deployed cloud recording server. Have a look at this guide to learn how to set up your [cloud recording server](https://github.com/AgoraIO-Community/Cloud-Recording-Golang)
+  final String? cloudRecordingUrl;
+
   /// (Optional) RTC Token value generated from the Agora dashboard.
   final String? tempToken;
 
@@ -33,6 +36,10 @@ class AgoraConnectionData {
   /// Whether you want to enable RTM or not. Enabling RTM adds the host controls which helps you to request a remote user to mute/unmute their video/mic. Host Controls are enabled by default, set this to `false` to disable it.
   final bool rtmEnabled;
 
+  final int? screenSharingUid;
+
+  final bool screenSharingEnabled;
+
   AgoraConnectionData({
     required this.appId,
     required this.channelName,
@@ -41,10 +48,13 @@ class AgoraConnectionData {
     this.rtmUid,
     this.username,
     this.tokenUrl,
+    this.cloudRecordingUrl,
     this.tempToken,
     this.tempRtmToken,
     this.areaCode = const [AreaCode.areaCodeGlob],
     this.rtmEnabled = true,
+    this.screenSharingUid = 1000,
+    this.screenSharingEnabled = true,
   });
 
   AgoraConnectionData copyWith({
@@ -57,8 +67,11 @@ class AgoraConnectionData {
     String? tempToken,
     String? tempRtmToken,
     String? tokenUrl,
+    String? cloudRecordingUrl,
     List<AreaCode>? areaCode,
     bool? rtmEnabled,
+    int? screenSharingUid,
+    bool? screenSharingEnabled,
   }) {
     return AgoraConnectionData(
       appId: appId ?? this.appId,
@@ -70,8 +83,11 @@ class AgoraConnectionData {
       tempToken: tempToken ?? this.tempToken,
       tempRtmToken: tempRtmToken ?? this.tempRtmToken,
       tokenUrl: tokenUrl ?? this.tokenUrl,
+      cloudRecordingUrl: cloudRecordingUrl ?? this.cloudRecordingUrl,
       areaCode: areaCode ?? this.areaCode,
       rtmEnabled: rtmEnabled ?? this.rtmEnabled,
+      screenSharingUid: screenSharingUid ?? this.screenSharingUid,
+      screenSharingEnabled: screenSharingEnabled ?? this.screenSharingEnabled,
     );
   }
 }